app, libgimp, pdb: use 64-bit window id
This is a follow-up of the clang issues signaled in #8649. When testing
CLANG64 on Windows under MSYS2, there was another similar issue in
gimpwidgets-utils.c. By changing the signature here we kind of snowball
into multiple functions using guint32 for window ids.

Here we change all window id usage to guint64. As said in a021b199
nowadays we need a 64-bit id for 64-bit OS's.